MNES-1063: Refactoring of AASListView component (#17)
Main refactored elements:

- AASListView component was split into smaller components
- AasListComparisonHeader - Header on the top of the page to visualize
selected AAS from the list
      - SelectProductType - Dropdown menu to select filter product class
      - AasList - Aas List 
      - AasListTableRow - Entry row for one AAS in the table
- ImageWithFallback - component ensures that a fallback icon (ShellIcon)
is displayed if the image specified by src fails to load.
      
Small refactoring to fix build.
