+++ src/app/[locale]/list/_components/AasListHeader.tsx
@@ -1,0 +1,12 @@
+﻿import { getTranslations } from 'next-intl/server';
+import { Typography } from '@mui/material';
+
+export default async function AasListHeader() {
+    const t = await getTranslations('aas-list');
+
+    return (
+        <Typography variant="h2" textAlign="left" marginBottom={2}>
+            {t('header')}
+        </Typography>
+    );
+}
