+++ src/app/[locale]/list/page.tsx
@@ -1,16 +1,13 @@
-import { Box, Typography } from '@mui/material';
-import { AASListView } from './_components/AASListView';
-import { getTranslations } from 'next-intl/server';
+import { Box } from '@mui/material';
+import { AasListView } from 'app/[locale]/list/_components/AasListView';
+import AasListHeader from 'app/[locale]/list/_components/AasListHeader';
 
-export default async function page() {
-    const t = await getTranslations('aas-list');
+export default function Page() {
     return (
         <Box display="flex" flexDirection="column" marginTop="20px" marginBottom="50px" width="100%">
             <Box width="90%" margin="auto">
-                <Typography variant="h2" textAlign="left" marginBottom={2}>
-                    {t('header')}
-                </Typography>
-                <AASListView />
+                <AasListHeader />
+                <AasListView />
             </Box>
         </Box>
     );
